A package to return colour gradients from CPTCITY

Colour palletes comes from http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html Rhw function cpt has two arguments n for the numbers and pal for the name or number of the pallete:

Details

"inferno"

"temperature"

"sst"

"temp"

"arendarl_arctic"

"arendarl_temperature"

"arendarl_zambezi_focus"

"arendarl_zambezi_proximity"

The palletes are ones of the most popular.

The palletes are available here: http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html