Function to return colour palletes functions from 'cpt-city'

This function return a collor pallete based on the name or position of the pallete.

cpt(pal = "mpl_inferno", n = 100)

Arguments

pal
Pallete of colors available or the number of the position

n
integer; number of colors of the same

Value

A colour pallete function.

Details

The cpt-city web archive comes from: http://soliton.vm.bytemark.co.uk/pub/cpt-city/index.html

Examples

{
library(cptcity)
image(matrix(1:100), col = cpt(pal = "mpl_inferno"))
find_cpt("temperature")
image(matrix(1:100), col = cpt("idv_temperature"))

#> Error: <text>:7:0: unexpected end of input
#> 5: image(matrix(1:100), col = cpt("idv_temperature"))
#> 6: 
#>   ^
# NOT RUN {
# Do not run
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions)) +
geom_raster(aes(fill = density))

 ggplot(faithfuld, aes(waiting, eruptions)) +
 geom_raster(aes(fill = density)) +
 scale_fill_gradientn(colours = cpt(n = 100))
# }
}

#> Error: <text>:1:1: unexpected '}'
#> 1: }
#>     ^